Add files via upload
This is the latest version of my python calculator. This calculator is aimed at primary school children, allowing them to do basic calculations using (+,-,*,/), as well as some more options for more advanced students, such as to the power of, square root, and squared. 

So far, the calculator does not have a UI and relies on terminal input and output. You can run it using the terminal in Visual Studio Code or any other terminal that has Python installed.